# thinx-api-openapi.yaml
openapi: 3.0.3
info:
  title: THiNX Device Management API
  description: |
    THiNX Device Management API provides functionality for managing IoT devices,
    firmware updates, and related services. The API supports both v1 and v2 endpoints.
  version: 2.0.0
  contact:
    name: THiNX Support
    url: https://thinx.cloud
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://rtm.thinx.cloud/api
    description: Legacy API (v1)
  - url: https://rtm.thinx.cloud/api/v2
    description: Current API (v2)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    
    sessionCookie:
      type: apiKey
      in: cookie
      name: x-thx-core
      description: Session cookie obtained after login
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authentication
      description: API key for device authentication
    
    ownerApiKeyAuth:
      type: object
      description: Owner ID and API key in request body
      properties:
        owner:
          type: string
          description: Owner ID
        api_key:
          type: string
          description: API key
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        status:
          type: string
          description: Error message
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: User's username or email
        password:
          type: string
          description: User's password
        remember:
          type: integer
          description: Set to 1 to extend session duration
          enum: [0, 1]
          default: 0
    TokenLoginRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: OAuth token
    LoginResponse:
      type: object
      properties:
        status:
          type: string
          example: "OK"
        success:
          type: boolean
          example: true
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: JWT refresh token
        redirectURL:
          type: string
          description: URL to redirect to after login
...
  parameters:
    XsrfTokenHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: |
        Anti-CSRF token. Must equal the `XSRF-TOKEN` cookie sent with the same request
        (compared in constant time) and must be bound to the `x-thx-core` session cookie sent
        with it. Obtain it by calling `GET /csrf-token` first and keeping every cookie that
        call sets (`x-thx-core` and `XSRF-TOKEN`) in a cookie jar. The token rotates at login:
        take the new value from the login response's `Set-Cookie`. Browser consoles do this
        automatically. Bearer (JWT) calls, and API-key calls that send no session cookie, are
        exempt and need neither the header nor the cookies. An API key never exempts the login,
        registration or password routes.
      schema:
        type: string
    XsrfTokenCookie:
      name: XSRF-TOKEN
      in: cookie
      required: true
      description: |
        Anti-CSRF cookie set by the API (`SameSite=Lax`, not httpOnly, `Path=/`), issued
        together with the httpOnly `x-thx-core` session cookie by `GET /csrf-token`. Its value
        must be echoed in the `X-XSRF-TOKEN` header, and it is only valid with the session it
        was issued for, so keep both cookies.
      schema:
        type: string
...
paths:
...
  /login:
    post:
      summary: User login
      description: |
        Authenticate a user with username and password or OAuth token.
        Requires the anti-CSRF double-submit pair (`XSRF-TOKEN` cookie + matching
        `X-XSRF-TOKEN` header, see `GET /csrf-token`); without it the call returns
        403 `csrf_token_invalid`.
      tags:
        - Authentication
      parameters:
        - $ref: '#/components/parameters/XsrfTokenHeader'
        - $ref: '#/components/parameters/XsrfTokenCookie'
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/LoginRequest'
                - $ref: '#/components/schemas/TokenLoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '403':
          description: |
            Authentication failed, or the anti-CSRF check failed. A CSRF failure has the body
            `{"success":false,"response":"csrf_token_invalid"}` (see
            `#/components/responses/CsrfTokenInvalid`); re-fetch `GET /csrf-token` and retry once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /logout:
    get:
      summary: User logout
      description: Destroy user session
      tags:
        - Authentication
      security:
        - sessionCookie: []
      responses:
        '302':
          description: Redirect to login page

  # Device endpoints
...